Explicitly cast argument to int

+++ vm/oop.hpp
@@ -282,7 +282,7 @@ Object* const cUndef = reinterpret_cast<Object*>(0x22L);
     }
 
     bool marked_p(unsigned int which) const {
-      return mark_ == which;
+      return mark_ == (int)which;
     }
 
     void mark(ObjectMemory* om, unsigned int which) {
